+++ src/main/java/site/moamoa/backend/service/module/member/MemberModuleServiceImpl.java
@@ -7,6 +7,8 @@ import site.moamoa.backend.api_payload.exception.handler.MemberHandler;
 import site.moamoa.backend.domain.Member;
 import site.moamoa.backend.repository.member.MemberRepository;
 
+import java.util.Optional;
+
 @Service
 @RequiredArgsConstructor
 public class MemberModuleServiceImpl implements MemberModuleService{
@@ -17,5 +19,15 @@ public class MemberModuleServiceImpl implements MemberModuleService{
     public Member findMemberById(Long memberId) {
         return memberRepository.findById(memberId)
                 .orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
+    }
+
+    @Override
+    public Optional<Member> findMemberBySocialId(String socialId) {
+        return memberRepository.findBySocialId(socialId);
+    }
+
+    @Override
+    public Member save(Member member) {
+        return memberRepository.save(member);
     }
 }
